Transfer appointments link, logout button from dashboard into sidebar
